+++ Tests/Runtime/CategoryFilteringTests.cs
@@ -29,14 +29,26 @@ namespace TraceForge.Tests
         [Test]
         public void CategoryVerbosity_CanBeSetHigherThanGlobal()
         {
-            TF.SetMinVerbosity(Verbosity.Trace);
+            TF.SetMinVerbosity(Verbosity.Info);
             TF.SetCategoryVerbosity(Categories.Network, Verbosity.Error);
 
-            TF.Debug(Categories.Network, "filtered debug");
+            TF.Info(Categories.Network, "filtered info");
             TF.Error(Categories.Network, "allowed error");
 
             Assert.AreEqual(1, _sink.Count);
             Assert.AreEqual(Verbosity.Error, _sink.Entries[0].Verbosity);
+        }
+
+        [Test]
+        public void CategoryVerbosity_CanBeSetLowerThanGlobal()
+        {
+            TF.SetMinVerbosity(Verbosity.Info);
+            TF.SetCategoryVerbosity(Categories.Network, Verbosity.Trace);
+
+            TF.Trace(Categories.Network, "allowed trace");
+
+            Assert.AreEqual(1, _sink.Count);
+            Assert.AreEqual(Verbosity.Trace, _sink.Entries[0].Verbosity);
         }
 
         [Test]
@@ -51,10 +63,13 @@ namespace TraceForge.Tests
         [Test]
         public void ClearCategoryVerbosity_RestoresGlobalFilter()
         {
-            TF.SetCategoryVerbosity(Categories.Network, Verbosity.Error);
+            TF.SetMinVerbosity(Verbosity.Info);
+            TF.SetCategoryVerbosity(Categories.Network, Verbosity.Trace);
+            Assert.IsTrue(TF.IsEnabled(Verbosity.Trace, Categories.Network));
             TF.ClearCategoryVerbosity(Categories.Network);
 
-            TF.Debug(Categories.Network, "now visible");
+            Assert.IsFalse(TF.IsEnabled(Verbosity.Trace, Categories.Network));
+            TF.Info(Categories.Network, "now visible");
             Assert.AreEqual(1, _sink.Count);
         }